requirements.txt
Pillow>=10.0.0
tqdm>=4.65.0
